Refactored JoinGatherer to new file and cleanup of ownership
